requirements.txt
cryptography>=42.0.0
Pillow>=10.0.0
qrcode>=7.4.2
PyYAML>=6.0

